update_2019-12-26--19:40

+++ commit.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-rm .*~
-rm *~
+if [[ -f .*~ || -f *~ ]] ; then
+  rm .*~
+  rm *~
+fi
 
 git add --all .
